Extract model creation
