+++ packages/backend/src/surveys/defaultSurveyFactory/dataModelFactory.js
@@ -1,0 +1,8 @@
+const { getSubmissionMetadata } = require("../sharedDataSchemas");
+const Util = require("../../utils/mongoose");
+
+module.exports = (survey, dataSchema) =>
+  Util.createModel(survey.collectionName, {
+    metadata: getSubmissionMetadata(survey),
+    surveyData: dataSchema,
+  });
