+++ packages/backend/src/surveys/gravedigger/model.js
@@ -1,0 +1,18 @@
+const { Surveys } = require("../config");
+const modelFactory = require("../defaultSurveyFactory/dataModelFactory");
+
+module.exports = modelFactory(Surveys.gravedigger, {
+  gravesite: {
+    type: String,
+    required: true,
+    index: true,
+  },
+  gravediggerPhoneNumber: {
+    type: String,
+    required: true,
+  },
+  burialsThatDay: {
+    type: Number,
+    required: true,
+  },
+});
